Created TUFRepository class
Moved some of the functions into the newely created class
TUFRepository. Having in class is convenient because it
is no longer necessary to pass into some of the functions.
Additionally, it eliminates the need to calculate certain
paths over and over again.
Replaced the function which returns a targets object with
a method which can return the appropriate object for each role.
Some smaller fixes