Removed AuthenticationRepo and TargetsRepo classes. Added targetsdb
- In order to make TAF as flexible as possible, removed AuthenticationRepo and TargetsRepo classes
added in the last commit. Kept the GitRepository class. The idea is that those who want to use TAF
will have to subclass that class.
- Instead of creating the targets inside the removed AuthenticationRepo class, they are created and
stored in a dictionary inside the targetsdb module. This implementation was inspired by TUF's roledb.
The idea is to have a dictionary, storing a dictionary for every authentication repository, and
inside that dictionary, a  dictionary per authentication repository's commit. Repositories can
be added and removed through time.