Updated files_to_keep, cli fixes
- In the Repository class there is now a private field which lists
all files which should remain targets even if not specified in data.
These files are required by the framework. Currently, the only such
file is repositories.json. The users can specify additional files which
should not be deleted if not present using an optional parameter of add_targets
- Smaller fixed - wrong package in setup.py, unused impoty, required
cli options changed to arguments.