Checking if metadata files remained the same
When refreshing metadata files, it is checked if metadata files
which should've remained the same based on the reference metadata
actually remained the same.
Added a cli command for updating a metadata files's expiration date.
This was useful for testing the previously mentioned functionality, but
I think that we'll need it later anyway.