Some fixes needed to get update working
Although some arguments were not needed by the git updater, they
were needed by the remote updater, so TUF's updater had to pass in
all those parameters. So, update could not be executed successfully.
Made changes to TUF, so that we don't have to use pylint disable.
Instead, kwargs are used when an aruments is needed by one and not
by the other handler.