Saving and loading last validated commit to/from a file
The last validated commit is saved to a file after the update is completed
successfully. It is loaded from a file when initializing commits in the git handler.
Reporting a more descriptive error if we cannot find commits since the last validated commit. That could indicate a force push removal of commits.
Checking if the last validated commit matched the repository's head commit and raising an error if that is not the case. Added some notes about what we could do in such a case.