Repo name fixes
Skipping validation of the repository's name if it is None. In such
case, the repo_path is set to root_dir and repo_name to the name
of the last directory on the root_dir path. These changes should
support the case when we only want to provide the repository's full path,
and have no need for validations.