WIP - making repo_name optional
If a repository's name should be validated, a class called NamedGitRepository
should be instantiated. This class extends GitRepository.
Deleted AuthenticationRepo class, to avoid having the same hierarchy of AuthenticationRepo classes. Moved the two methods contained by this class to the updater.
Updated the handler.

Still need to refactor the updater to support the case of providing the whole repo path and not the repository's name