Calling update with and without repository's name
The updater can be called both with and without providing name of the authentication repository which is supposed to be updated. Since TUF's updater expects to receive repository's name, in case when it is not provided, the repository's full path (which is expected to be provided in that case) is split into repo dir and repo name. A setting set in TAF is used to indicate if the repository's nam should be validated. This is the easiest way to implement what we want without modifying TUF's updater again