Setting repository name and path fixes (#24)
* Repo name fixes

Skipping validation of the repository's name if it is None. In such
case, the repo_path is set to root_dir and repo_name to the name
of the last directory on the root_dir path. These changes should
support the case when we only want to provide the repository's full path,
and have no need for validations.

* Update docstring

* Removed git repository's name property and bare attribute

- Name is no longer needed since repo_name will always store a repository's name
- Removed bare attribute. Bare is simply a parameter of the clone method.
- Updated GitRepository's docstring

* WIP - making repo_name optional

If a repository's name should be validated, a class called NamedGitRepository
should be instantiated. This class extends GitRepository.
Deleted AuthenticationRepo class, to avoid having the same hierarchy of AuthenticationRepo classes. Moved the two methods contained by this class to the updater.
Updated the handler.

Still need to refactor the updater to support the case of providing the whole repo path and not the repository's name

* Calling update with and without repository's name

The updater can be called both with and without providing name of the authentication repository which is supposed to be updated. Since TUF's updater expects to receive repository's name, in case when it is not provided, the repository's full path (which is expected to be provided in that case) is split into repo dir and repo name. A setting set in TAF is used to indicate if the repository's nam should be validated. This is the easiest way to implement what we want without modifying TUF's updater again

* Addressed CR comments