Updater - added support for multiple target repos branches
A target file corresponding to a target repository might contain name of the
target repository's branch. If that is the case,
we want to update that branch, and not the repository's master branch.
Additionally, if there is more than one new authentication repository commit,
it is possible that more than one branch of a target repository will need to
be updated.
If branch is not specified in the target file, it is assumed that the target to
updated.