Moved and updated add_signing_key and removed register_target_file
Removed register_target_file as it is not necessary and can be confusing. targets sign should be used instead
Moved add_signing_key to a subcommand. Updated it so that public key which corresponds to the private key that should be used for singing can be copy-pasted into the termial instead of having to save it to a file