Logging improvements
Lowered logging level in a few places. When an error occurs when executing a
git command and we neither want to reraise the error not use a customized log error message, such error is not that important. For example, a file does not exist at a revision. We often traverse through all commits and, e.g. try loading repositories.json. This file might not be available at the initial revision and that is not a problem.
Remove tuf logging when not needed
Small refactoring