Setting urls based on mirrors.json
Added support for specifying mirrors in a separate mirrors.json. Mirrors are templas, containing org_name and repo_name named arguments. Something like https://github.com/{org_name}/{repo_name}. It is assumed that names of repositories in repositories.json are in the org_name/repo_name format. When calculating urls, org_name and repo_name are replaced with actual names based on values in repositories.json. Added a test for this