Fixed how it is checked if there are additional commits
The idea is to raise an error if there are commits which have not beeen
pulled yet in case of a repository where unauthenticated commits are
allowed. We find all commits which are newers than the last authenticated
commit, but some of them might have already been pulled. So, remove
those which are older or equal to the current head.