Handle worktrees, allow specification of conf dir location
When merging commits, we need to check out branches which need to be
updated. That cannot be done if they are checked out in a different worktree.
So try to switch to that worktree and update the branch. Currently, there is
no validation if that is possible or permissible
Specify location of the conf directory through the settings file