WIP - work on reducing number of str Path conversions and instantiating repos with library_dir
The code revolving around halding different variations of names and paths
was getting to complicated, so introduced instantiation of repositories
using library_dir or name or path. If all 3 are defined, it is checked
if the calculated path is the same as the proveded one.
Some work done on minimizing str/Path conversions