add pytest11 endpoint only if tests are available
taf.tests module is not available in the wheel,
but the pytest11 entrypoint requires taf.tests, and
is always called by any pytest in any environment
that has taf installed. So any call to pytest will
always fail if taf wheel is installed.

By only conditionally adding pytest11 if taf.tests
exists, the entrypoint will only be available when
installed as source.