perf: implement get_file in pygit2
Git cli operations are slow on windows. Since git.py is just a wrapper around
git cli, it is also very slow.

By reimplementing get_file using pygit2 cumtime for that call drops from 524
secs to 3.59 secs.

This requires a new dependency on pygit2. pinned to a very old version that is
available on our old ubuntu boxes and has built wheels for 3.6.

Added a new PyGitRepository that hides the complexity of pygit2.

I recall that renata was having a lot of problems with pygit2 being flaky on
windows, so we always fall back to the git cli.

We must cleanup the pygit2 repo because it keeps file handles open, so can't
delete it or anything else.