refactor: move dependencies to `repositoriesdb`
Since auth_repo object has a dependencies field that will later
be returned by the updater, it can be a property.

* Added a setter for that dependencies property.

* Based on discussion found in PR #206, the logic for reading
`dependencies.json` into an attribute that an auth_repo will later use
is now moved to `repositoriesdb`.  Makes sense that we don't need to load
dependencies.json from each of the commits to set the `dependencies`
attribute, just the most recent commit. Because of that, we also don't
need to instatiate a different exception class, since it's
already wrapped up by `InvalidOrMissingMetadataError`. Removed the raise
error from that, to have feature parity with `_load_mirrors_json`.