feat: add unsafe flag to repositories (#229)
A security update made to git prevents execution of git commands if
the containing directory is owned by a different user. For example, if
repositories are owned by root and www-data tries to execute a git command.
It only affects commands executed using git cli - _git method of the git repo
class in TAF. pygit2 usage is not affected.
`allow_unsafe` flag can be set when instantiating the repositories allowing
the mentioned safety feature to be ignored. This is done by marking the
containing directory as safe. It will be set to False by default.

Signed-off-by: Renata <rvaderna@openlawlib.org>