perf: only update a list of roles if a metadata file was added
If a new delegation is added and the new role contains targets, there
will have to be a metadata file. If no such changes are detected,
we can be cofident that there are no new roles. No need to use the
repository_at_revision context manager just to get a list of target
roles for each commits