Performance and profiling (#228)
* feat: cache git files and profiling

- Added timed_run decorator to updater calls which prints the total
execution time
- Added profile flag to update cli. This makes it possible to easily
generate .prof files.
- Added a few comments which make it easier to understand integration with
the TUF updater explaining why that code should not be replaced with our own
optimized solutons

* chore: cleanup and formatting

* perf: only update a list of roles if a metadata file was added

If a new delegation is added and the new role contains targets, there
will have to be a metadata file. If no such changes are detected,
we can be cofident that there are no new roles. No need to use the
repository_at_revision context manager just to get a list of target
roles for each commits

* chore: update changelog and cleanup

* chore: remove unused attribute

* fix: do not raise an error if a metadata does not exist at revision (was removed)

Signed-off-by: Renata <rvaderna@openlawlib.org>