refact: improve updater performance
Use `verify_length_and_hashes` instead of downloading targets. Problem
with downloading targets or using `find_cached_target` is that TUF
expects files to exist on filesystem. Downloading targets means they
will be persisted, and finding cached target means that a persisted
target will be open. With that in mind, storing targets on a
filesystem for each revision (in a repoistory with potentially
thousands of revisions) will be a performance bottleneck.
Instead we call `verify_length_and_hashes` on a loaded targets
directly.

Clean up unused methods in GitUpdater.