fix: update `invalid-metadata` test repository
On TUF versions that were < 0.20.0 it was expected that metadata with a key
with incorrect signature was invalid. However, since we can verify that
delegated_metadata is signed by threshold of unique keys, it is possible
to have an invalid signature. That is, as long as there are >= threshold
amount of valid signatures, the metadata is considered valid.