fix: bring `updater_tests` up-to-date with 2.0.0 TUF updater
ReplayedMetadataError has been removed. Instead, error message is more
informative.

METADATA_CHANGED_BUT_SHOULDNT is a TAF related error. We used to raise
that error when `current` and `previous` metadata was the same. TUF does
not keep track of current and previous anymore, so this test scenario is not
relevant to latest updater

fix: remove unused updater repo