fix: remove all_fetched_commits which does not always work
Instead of all_fetched_commits, all_commits_on_branch, which
was implemented using pygit2 can be called - as long as the
branch which is passed to it is a remote branch.
all_fetched_commits would return an empty list in some cases when
there were commits on the remote branch. all_commits_on_branch returned
the correct commits in those cases on the other hand.