feature: ignore missing branch file in some cases
In some cases, branch file might not be present in the first new commit
of the authentication repository. check_branch_roles turned into a dictionary
where repo name is the key and a flag indicating if a missing branch
is a valid state or not in case of that repo. This should be made
more generic in the future