Branch check update - support first commits on branches with a missing branch file (#292)
* feature: ignore missing branch file in some cases

In some cases, branch file might not be present in the first new commit
of the authentication repository. check_branch_roles turned into a dictionary
where repo name is the key and a flag indicating if a missing branch
is a valid state or not in case of that repo. This should be made
more generic in the future

* fix: rework check_branch_id

A branch file can be inside a folder named after the role inside the
targets folder, but also directly inside targets, depending on
how many roles need these files. Only try to load a branch file
from a certain location if that location was delegated to the current role

* chore: formatting

* fix: fix capstone check

Check if role is allowed to sign a certain path before checking
if capstone exists at that location

* chore: update changelog