refact: move add_signin_key to roles cli, rework loading of keys
Attempt to load a key from a YubiKey without asking the user to insert
it. If the correct YubiKey is already inserted, the key will be loaded.
If that is not the case, try to load from the keystore location.
If the key is not there, as the user if they want to sign using a yubikey
and ask them to insert it
Work on removing roles_key_info