feat: add prompt_for_keys flag
Added prompt_for_keys flag to all functions that load signing keys.
If set to True, we'll ask the user if they want to manually enter
the key if not using yubikeys and it cannot be loaded from keystore files.
If set to False, the execution of the program will stop if not using Yubikeys
and the key is not found inside the provided keystore directory