feat: ask the user to insert yubikey if public key is provided if threshold is not reached
If public key of a yubikey is not provided, the user will be asked to insert
the yubikey since public key will need to get extracted too. If public key
is provided, the user will need to insert that yubikey only if there is an
insufficient number of signing keys. So, say there are 3 keys, user1, user2 and user3
where public keys of user1 and user2 are provided, while there is no user key
of user3. The user will need to insert user3 key. If threshold is 1, that
should be enough. Otherwise, they will be asked to insert user1 or user2 if
threshold is 2 and both is threshold is 3