feat chore: Added alias attribute to repositories, logging improvements
- At the start of the update process, an error message used to be shown
saying that the default branch could not be determined. There is not need
to attempt to determine the default branch if the repository is not yet cloned.
- During the update process, the updater clones the remote reposiory to
a temp folder and this temp folder's name used to be used when logging.
E.g. 327fshkewrew/law would get set as the repository's name. In order
to be able to reference a repository by a different name when logging,
added a new attribute called alias. So now we can reference this temp
repository by a more descriptive name.