fix: use provided url or get repomote url
`self.urls` is not deterministic between OSs, so on Windows it
returns remote repositories last, and on macOS it returns them first.
This commit makes it so we don't rely on `self.urls`.

Also, since the for loop doesn't access or use url variable, I've
removed it.

See issue: #372 for details.