fix: remove `last_validated_commit` entry
We now store the exact same commit hash for the authentication repository
in the `<org>/<name>` format. So seems like an unnecessary entry in the file.
Ran the updater and the update looks good and checked the `last_validated_commits` file.

Seems like we only relied on `last_validated_commit` entry in our testing framework,
so tried updating it in place.