fix: get top commit from remote branch first
On our CI our repos don't pull down remote tracking branches as local branches. But because
the repositories are not bare, we resort to checking the local branch (without remotes).
The new approach in this commit is to get the top commit off of remote branch, then if that fails resort to checks whether the repo is bare/or not.