fix: avoid nesting 'Commitish' objects
It can sometimes happen that we unintentionally convert an already
built Commitish obj with 'from_hash', which then returns
Commitish(Commitish(hash)). This leads to some pretty difficult to debug
situations. So to avoid that problem altogether, we can simply
return the original hash, if the class is already set up.