fix: make 'target_repos' optional again
While the original fix did resolve the default branch issue, we now
force the user to always pass in 'target_repos'. This doesn't make much
sense in situations where we don't need target repos dict (such as publish server).

To avoid unnecessarily passing in target_repos each time, set it to be optional
(as was the case before), but do rely on figuring out the default branch if it is present.