fix: sort order of roles to load
Previously we used to iterate over a `set` of roles, then load
yubikeys for a given role. However, this lead to non-deterministic
yubikey prompts, since sometimes the first role that gets loaded is `targets`
instead of `root`. To resolve, we always sort by role, with root being
the first one to get prompted (if exists).