feat(test): add test case for signing on windows
Issues crop up when signing on windows due to target files getting
auto-converted to CRLF line endings. This caused our hash file compare
to break on Windows in cases where we use LF. To resolve, we updated the
hash file compare function to convert to LF. This test case shows that
only currently signed file metadata gets updated, and not any additional
metadata (even when converting it to CRLF prior).