fix: unauthenticated commits list should always contain all commits after last validated
During the merge step, the latest branch is reset to the last validated
commit. If the branch was on the latest commit (and that commit is
not authenticated), setting the list of authenticated commits to just commits
following the top of the branch results in an empty list. So when the
branch is reset to the last validated commit, we incorrectly return an empty
list.